requirements.txt
fastapi==0.95.2
uvicorn==0.21.1
python-dotenv==1.0.0
httpx==0.24.1
pydantic==1.10.7
python-multipart==0.0.6
requests==2.31.0